Moved reducer method to Certification.js to be used by other pages.
